pyproject.toml
[tool.poetry]
name = "littlerpg-discord-bot"
version = "0.1.0"
description = ""
authors = ["RRR-09 <79768142+RRR-09@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
python-dotenv = "^0.19.2"
"discord.py" = "^1.7.3"
six = "1.14.0"  # needed for mcstatus
mcstatus = "^7.0.0"
pytz = "^2021.3"
requests = "^2.26.0"
PyYAML = "^6.0"
mcipc = "^2.3.3"
pysftp = "^0.2.9"
parse = "^1.19.0"
mctools = "^1.1.2"
fastapi = "^0.74.1"
uvicorn = {extras = ["standard"], version = "^0.17.5"}

[tool.poetry.dev-dependencies]
black = "^21.12b0"
bandit = "^1.7.1"
flake8 = "^4.0.1"
isort = "^5.10.1"
mypy = "^0.920"
ossaudit = "^0.5.0"
types-PyYAML = "^6.0.1"
types-pytz = "^2021.3.3"
types-requests = "^2.26.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# Customizations
[tool.isort]
src_paths=["src"]
profile = "black"

[tool.mypy]
python_version = "3.9"

[[tool.mypy.overrides]]
module = ["discord.*", "mctools.*", "parse.*", "cogs.*"]
ignore_missing_imports = true
